+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- [#534](https://github.com/teambit/bit/issues/534) auto-tagging feature updates the components impacted by auto-tagging
+
 ## [0.12.0-ext.19] - 2018-01-11
 
 - fix #595, allow one dot in the scope name